fix: 認証メール送信時にメールアドレスの確認

+++ milk/lib/milk_web/controllers/conf_num_controller.ex
@@ -4,31 +4,36 @@ defmodule MilkWeb.ConfNumController do
   import Bamboo.Email
   alias Milk.ConfNum
   alias Milk.Email.Auth
+  alias Milk.Accounts
 
   @doc """
   Send an email for verification.
   FIXME: emailのexist確認
   """
   def send_email(conn, %{"email" => email}) do
-    number =
-      :rand.uniform(9999)
-      |> Integer.to_string()
-      |> String.pad_leading(4, "0")
+    exists? = Accounts.is_email_exists?(email)
 
-    Task.start_link(fn -> 
-      new_email(
-        to: email,
-        from: "adhisuabeba@gmail.com",
-        subject: "confirmation number",
-        text_body: number
-      )
-      |> Milk.Mailer.deliver_now
-      
-      ConfNum.delete_conf_num(email)
-      ConfNum.set_conf_num(%{email => number})
-    end)
+    if exists? do
+      number =
+        :rand.uniform(9999)
+        |> Integer.to_string()
+        |> String.pad_leading(4, "0")
 
-    json(conn, %{result: true})
+      Task.start_link(fn -> 
+        new_email(
+          to: email,
+          from: "adhisuabeba@gmail.com",
+          subject: "confirmation number",
+          text_body: number
+        )
+        |> Milk.Mailer.deliver_now
+        
+        ConfNum.delete_conf_num(email)
+        ConfNum.set_conf_num(%{email => number})
+      end)
+    end
+
+    json(conn, %{result: exists?})
   end
 
   @doc """
